refactor(cardLogo): 카드 프리뷰에서 카드 로고 컴포넌트 분리
